// src/types/comingData/devices.ts
import { z } from "zod";

export type DeviceApplicationT = {
  deviceId: number;
  serialNumber: string;
  model: string;
  packetType: "STD" | "ACC";
  suffix: string | null;
  devicePhoneNumber: string;
  isConnectedVehicles: boolean;
  companyApplicationId: string;
};




export const DeviceApplicationSchema = z.object({
  // Use .optional() if this schema is used for "Create" forms 
  // where the ID is generated by the database later.
  deviceId: z.number().int().positive(),

  serialNumber: z.string().min(1, "Serial number is required"),

  model: z.string().min(1, "Model is required"),

  // z.enum is cleaner for "STD" vs "ACC"
  packetType: z.enum(["STD", "ACC"], {
    error: () => {
      return 'Packet Type is Required'
    }
  }),


  suffix: z.string().nullable().optional(),

  isConnectedVehicles: z.boolean().default(false),

  // Adding a regex or min length for phone numbers is good practice
  devicePhoneNumber: z.string().min(11, "Invalid phone number"),

  companyApplicationId: z.string()
});

// Extract the Type for use in your components
export type DeviceApplicationSchemaT = z.infer<typeof DeviceApplicationSchema>;